address comments and update notice
