+++ create-notice.sh
@@ -54,6 +54,7 @@ function main {
     add_license "ijson" "https://raw.githubusercontent.com/ICRAR/ijson/master/LICENSE.txt"
     add_license "google-resumable-media" "https://raw.githubusercontent.com/googleapis/google-resumable-media-python/master/LICENSE"
     add_license "google-auth" "https://raw.githubusercontent.com/googleapis/google-auth-library-python/master/LICENSE"
+    add_license "aiokafka" "https://raw.githubusercontent.com/aio-libs/aiokafka/master/LICENSE"
 
     # transitive dependencies
     # Jinja2 dependencies
